Fix mistake in class type of ArffFiles Add some type casting to CPPFImdlp Add additional path to datasets in tests Fix some smells in sample Join CMakeLists

+++ CPPFImdlp.cpp
@@ -3,7 +3,6 @@
 #include <set>
 #include <cmath>
 #include <limits>
-#include <cmath>
 #include "CPPFImdlp.h"
 #include "Metrics.h"
 namespace mdlp {
@@ -21,7 +20,7 @@ namespace mdlp {
         if (proposed_cuts == 0) {
             return  numeric_limits<size_t>::max();
         }
-        if (proposed_cuts < 0 || proposed_cuts > X.size()) {
+        if (proposed_cuts < 0 || proposed_cuts > static_cast<float>(X.size())) {
             throw invalid_argument("wrong proposed num_cuts value");
         }
         if (proposed_cuts < 1)
@@ -125,8 +124,8 @@ namespace mdlp {
             // Cutpoints are always on boundaries (definition 2)
             if (y[indices[idx]] == y[indices[idx - 1]])
                 continue;
-            entropy_left = precision_t(idx - start) / elements * metrics.entropy(start, idx);
-            entropy_right = precision_t(end - idx) / elements * metrics.entropy(idx, end);
+            entropy_left = precision_t(idx - start) / static_cast<float>(elements) * metrics.entropy(start, idx);
+            entropy_right = precision_t(end - idx) / static_cast<float>(elements) * metrics.entropy(idx, end);
             if (entropy_left + entropy_right < minEntropy) {
                 minEntropy = entropy_left + entropy_right;
                 candidate = idx;
@@ -148,8 +147,8 @@ namespace mdlp {
         ent1 = metrics.entropy(start, cut);
         ent2 = metrics.entropy(cut, end);
         ig = metrics.informationGain(start, cut, end);
-        delta = log2(pow(3, precision_t(k)) - 2) -
-            (precision_t(k) * ent - precision_t(k1) * ent1 - precision_t(k2) * ent2);
+        delta = static_cast<float>(log2(pow(3, precision_t(k)) - 2) -
+            (precision_t(k) * ent - precision_t(k1) * ent1 - precision_t(k2) * ent2));
         precision_t term = 1 / N * (log2(N - 1) + delta);
         return ig > term;
     }
